Various changes that get the script to run
Correct PCT_NATVEG name. Check that landmasks are the same
for the glacier, PFT, and lanwat files. Turn off processing
of soitex, as it's not modefied. Add LATIXY/LONGXY to soil color file.
Changes from Michele P. to get nn_fill to work. Turn on the
section for new points, and add the needed do loops to it.